+++ libtf/ft_strlcat.c
@@ -6,15 +6,15 @@
 /*   By: alfrsanc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/03 18:45:35 by alfrsanc          #+#    #+#             */
-/*   Updated: 2025/10/08 11:57:26 by alfrsanc         ###   ########.fr       */
+/*   Updated: 2025/10/10 19:08:57 by alfrsanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	*ft_strlcat(char *dst, char *src, unsigned int dsize)
+unsigned int	*ft_strlcat(char *dst, char *src, size_t dsize)
 {
-	unsigned int	dstlen;
-	unsigned int	srclen;
-	unsigned int	i;
+	size_t	dstlen;
+	size_t	srclen;
+	size_t	i;
 
 	dstlen = 0;
 	srclen = 0;
